requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
typing-extensions>=4.8.0
pydantic>=2.5.0
